jobs/webcompat-kb/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["webcompat_kb"]

[project]
name = "webcompat-kb"
description = "Import webcompat data into BigQuery"
authors = [
  {name = "Mozilla Corporation"}
]
version = "0.1.0"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
license = "MPL-2.0"
dependencies = [
  "bugdantic==0.2.6",
  "google-cloud-bigquery==3.34.0",
  "httpx[http2]>=0.28.1",
  "jinja2>=3.1.6",
  "pydantic==2.11.5",
  "tomli-w>=1.2.0",
  "web-features==3.0.0",
]

[project.optional-dependencies]
test = [
  "mypy==1.16.0",
  "ty==0.0.15",
  "pandas-stubs==2.2.2.240603",
  "pytest-ruff==0.4.1",
  "pytest==8.4.0",
  "ruff==0.11.13",
  "types-cachetools==6.0.0.20250525",
  "types-cffi==1.17.0.20250523",
  "types-protobuf==6.30.2.20250516",
  "types-pyOpenSSL==24.1.0.20240722",
  "types-python-dateutil==2.9.0.20250516",
  "types-pytz==2025.2.0.20250516",
  "types-requests==2.32.0.20250602",
  "types-setuptools==80.9.0.20250529",
  "types-tqdm==4.67.0.20250516",
]

[project.scripts]
webcompat-etl = "webcompat_kb.main:main"
webcompat-add-metric = "webcompat_kb.commands.add_metric:main"
webcompat-backfill-history = "webcompat_kb.utils:backfill_history"
webcompat-backfill-metric = "webcompat_kb.commands.backfill_metric:main"
webcompat-check-templates = "webcompat_kb.commands.checkdata:main"
webcompat-metric-rescore = "webcompat_kb.commands.metric_rescore:main"
webcompat-render = "webcompat_kb.commands.render:main"
webcompat-update-staging-data = "webcompat_kb.commands.update_staging_data:main"
webcompat-site-rank-update = "webcompat_kb.commands.site_rank_update:main"
webcompat-validate = "webcompat_kb.commands.validate:main"


[tool.pytest]
testpaths = ["tests"]

[tool.mypy]
disallow_untyped_defs = true
